256 microseconds delay inserted before actual ID byte transmission.
From the CCD Bus User's Manual:

Arbitration may begin either when bus idle begins or at a later time when a user microprocessor begins transmitting a new message. A 256 microseconds delay between the beginning of bus idle, (i.e. IDLE-going low), and the start of a new message, (i.e. the falling edge of the start bit), is called the start of message delay (SOM Delay).

[...]

The CCD-bus system contains a requirement that modules delay the beginning of each message until 256 microseconds after the bus goes idle. However, this was not enforced by the CCD-I hardware in SCI mode. The CCD-II chip does enforce this requirement.